more test wrangling for permissions
